Factor PayeeTableRow out to component file
